Format params
I received complaints from the user that despite the status 200, he was not receiving the alert, I tested it and found the problem.

to fix this, considering a GET call, in params I modified the metadata format and removed the status, after that alerts in sequence are being received.

For some reason when the parameter Status is sent, only one request is received and in a sporadic way.

Besides adding urllib3 to remove warnings.