update webhooks
File: webhook_reader_flask.py

I made a small modification to the file to make it easier for the client to start the script with just a basic command without the need to export environment variables.
Fixed the handling of the file when no image is sent through the webhook.
File: webhook_reader.py

Fixed saving the upload with the pattern sent by the request.
Fixed the handling of the file when no image is sent through the webhook.
File: webhook_reader.js

Fixed the filename field in the request.
Fixed the handling of the file when no image is sent through the webhook.